+++ CHANGELOG.md
@@ -73,3 +73,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.5] - 2024-7-10
 
 ### Added
+- My Profile page