+++ CHANGELOG.md
@@ -75,3 +75,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - My Profile page
 -- It will not allow a non-signed in user to access the page
+-- The user can change their Username, Password(if they confirm it first), Email, Phone Number, Full Name, Address. They can check themselves in or out of being a Volunteer or Requestor.